Fix parameterized explicit providers (#435)
This change fixes `ProviderResource`, which wasn't passing along the
parameterization information to its base class.

This should definitely have a test. It will be covered by the
`l2-explicit-parameterized-provider` conformance test, but that is
currently blocked on https://github.com/pulumi/pulumi-dotnet/issues/432
(which is imminent, but I'd prefer not to hold up the fix for this on
that). In the meantime, I've tested this manually and confirmed it fixed
the problem in the issue.

Fixes #424