[language-runtime] Implement GetRequiredPackages to replace GetRequiredPlugins (#440)
### Description

The .NET language runtime currently uses `GetRequiredPlugins` to
determine which resource plugins a program needs in order to run.
However, `GetRequiredPlugins` doesn't account for the parameterization
of the plugins. This is where `GetRequiredPackages` comes into play
which allows the language runtime to return required plugins/packages
along with their parameterization info (loaded from `pulumi-plugin.json`
files).

This also allows the language runtime to correctly work with `pulumi
import` for parameterized providers.

Fixes #432 and unskips tests:
 - l2-parameterized-resource
 - l2-explicit-parameterized-provider

> Note: the tests were failing locally but that was because the local
package nugets used in the conformance tests had old contents and needed
to be invalidated. This so that `pulumi-plugin.json` could be filled
with the parameterization values.