Avoid calling invokes with dependencies on unknown resources (#441)
.NET implementation of https://github.com/pulumi/pulumi/pull/18133

DependsOn for resources is an ordering constraint for register resource
calls. If a resource R1 depends on a resource R2, the register resource
call for R2 will happen after R1. This is ensured by awaiting the URN
for each resource dependency before calling register resource.

For invokes, this causes a problem when running under preview. During
preview, register resource immediately returns with the URN, however
this does not tell us if the resource "exists".

Instead of waiting for the dependency's URN, we wait for the ID. This
tells us that whether a physical resource exists (if the state is in
sync), and we can avoid calling the invoke when it is unknown.

The following example fails without this change:

```csharp
using Pulumi;
using Pulumi.Gcp.Organizations;
using Pulumi.Gcp.Compute;
using System.Collections.Generic;

return await Deployment.RunAsync(async () =>
{
    var config = new Config();
    var billingAccountId = config.Require("billing-account");

    var billingAccount = await GetBillingAccount.InvokeAsync(new GetBillingAccountArgs
    {
        BillingAccount = billingAccountId
    });

    var project = new Project("project", new ProjectArgs
    {
        BillingAccount = billingAccount.Id,
        Name = "project-nodejs",
        AutoCreateNetwork = false,
        DeletionPolicy = "DELETE",
    });

    var zones = GetZones.Invoke(new GetZonesInvokeArgs
    {
        Project = project.ProjectId,
        Region = "us-central1"
    }, new InvokeOutputOptions
    {
        DependsOn = { project }
    });

    return new Dictionary<string, object?>
    {
        ["zones"] = zones,
    };
});
```