Fix dotnet version (#445)
With the update to Ubuntu 24.04 for the runners, the preinstalled .NET
versions have changed. This revealed some issues with how we select the
.NET version to use in the matrix tests. The `actions/setup-dotnet`
action does install the version we want, but the dotnet command still
use the runner’s default version.

We have to create a `global.json` to make it use the desired version.


https://github.com/actions/setup-dotnet?tab=readme-ov-file#matrix-testing