Don't parse runtime options at startup (#451)
A more general fix to
https://github.com/pulumi/pulumi-dotnet/issues/443.

When starting up the dotnet language host don't try to parse the runtime
options, and especially don't error due to missing "dotnet". It's valid
to start the host in modes that "dotnet" won't be needed, even though
the runtime options might not be clear about that.

In the future the engine won't even send the runtime options at startup,
and given that multiple plugins can use the same language host it's
incorrect to respect just the first runtime options seen anyway.