Fix adding duplicate key-value pairs to InputMap (#462)
Fixes https://github.com/pulumi/pulumi-dotnet/issues/458

Before https://github.com/pulumi/pulumi-dotnet/pull/449 it used to be
possible to `Add` the same key value pair to `InputMap` multiple times.
This because of a feature of the underlying ConcurrentDictionary which
will use `EqualityComparer` to see if the key and value being added are
the same as what's in the dictionary already and if so just returns the
existing dictionary.

#449 broke this because now all the values of the dictionary are
`Input`s which don't have a sensible notion of equality. This adds some
extra handling in our Add method to try and maintain this feature.

It _does_ have two oddities:
1. The exception is recorded on the value, not the map overall. This
shouldn't be observable because users can only resolve InputMap as a
single `Input<Dict<K, V>>` and not directly access the nested input
values. But it's worth calling out to keep in mind for the future.
2. The value always merges the secretes and dependencies of all `Input`s
added to the one key. Those properties don't get overwritten, so if you
add `("k", secret("hello"))` then add `("k", "hello")` you might expect
that value to no longer be secret because it's been overwritten, but in
fact the secretness is merged and so it's still secret. Fixing this
would require a custom apply implementation to await both values, then
_iff_ they're both known keep just the output data from the later one,
but you'd still end up with merged data if either was unknown. All in
all it didn't feel worth the complexity to do different here.