Handle null runtime values in InputMap/List cast (#459)
Fixes https://github.com/pulumi/pulumi-dotnet/issues/456

This is a bit of a sad fix but `InputMap` used to tolerate the implicit
cast from dictionary being called with null by just passing that null
down. If you tried to call Add or enumerate it would NRE, but if it just
got assigned to another resource property the serialiser could handle
it.

https://github.com/pulumi/pulumi-dotnet/pull/449 broke this because it
always tried to enumerate the incoming immutable dictionary to build the
new internal nested immutable dictionary. This fixes it to behave as
before were null flows through as null.