Add path.root (rootDirectory) equivalent to pcl and generate for go/python/node (#18595)
Reference:
https://developer.hashicorp.com/terraform/language/expressions/references#filesystem-and-workspace-info

pcl.root as per the docs returns the project root, which we define as
the
location of the Pulumi.yaml (for example if we pass in the `-C`/`cwd`
parameter to execute
in another directory).

path.cwd is the directory where terraform is run from, this means even
if
`-C`/`cwd` are set, it will return the current directory. This is not
included in this PR and will require a bit more (but similar) work.
(sending CWD through rpcs to language, and all the piping from this CL.
etc).

To do after this submits:
Make similar changes in pulumi/pulumi-dotnet and pulumi/pulumi-java

Part of #18565