Consistent inheritance behavior for InputAttribute and OutputAttribute (#506)
The InputAttribute and OutputAttribute have an AttributeUseage with
inherit: true.

When using abstract classes and properties, the current implementation
does not handle inheritance of the attributes consistently. In some
cases, the attributes are checked with inherit: false, in other cases
with inherit:true.

I have made the behavior consistent and as expected from the
AttributeUseage of InputAttribute and OutputAttribute.