Specify StringComparison for correctness
This change specifies a StringComparison to calls to `StartsWith` and `EndsWith`, and enables the analyzer rule to enforce it going forward.

Note: We should enable a more aggressive set of analysis rules (likely via `<AnalysisMode>Recommended</AnalysisMode>` in the .csproj files), but that requires more fixes, so just enabling this specific rule for now.