Capture async context for engine callback methods
The `Callback` system is what allows the engine to callback to the
langauge runtime to invoke generic code. Currnently just used for the
transforms system, but soon to be extended for other uses as well.

We we're incorrectly not propogating the async context from where the
callback system was created (somewhere inside the users program as part
of resource registration call) to the threads that the callbacks run on
(in the kestral server threadpool).

This meant that inside a transform function you couldn't access any of
the async context that you had access to just outside the transform
function. Importantly this includes the deployment instance that allows
access to the engine.

We probably wouldn't recomend registering more resources from the
transform function (although, you _could_) but users definitly should
have access to the engine logging system from inside the transform
function.

This PR adds the requiste ExecutionContext capture and run calls to copy
and restore the async context from the main program to the threads
running the callback functions.