Capture async context for engine callback methods (#561)
The `Callback` system is what allows the engine to callback to the
language runtime to invoke generic code. Currently just used for the
transforms system, but soon to be extended for other uses as well.

We were incorrectly not propagating the async context from where the
callback system was created (somewhere inside the users program as part
of resource registration call) to the threads that the callbacks run on
(in the kestral server threadpool).

This meant that inside a transform function you couldn't access any of
the async context that you had access to just outside the transform
function. Importantly this includes the deployment instance that allows
access to the engine.

We probably wouldn't recommend registering more resources from the
transform function (although, you _could_) but users definitely should
have access to the engine logging system from inside the transform
function.

This PR adds the requisite ExecutionContext capture and run calls to
copy and restore the async context from the main program to the threads
running the callback functions.

---------

Co-authored-by: Justin Van Patten <jvp@justinvp.com>