Use standard naming for generic type parameters
This change modifies all non-standard generic type parameter names to use the standard convention of being prefixed by `T`, making them idiomatic for .NET. It avoids having to disable the CA1715 analysis rule in a bunch of places. This is a non-breaking change (unless you are doing something highly unusual with reflection).