Fix function call lifting (#19187)
Fixes https://github.com/pulumi/pulumi/issues/8439

This moves the output lifting that we were doing unconditionally on
functions to the function binding itself. That is if a function can
handle plain or output values (like toJSON) it handles that in its
signature binding. Otherwise we only lift the function if an argument is
an output but the parameter is plain. This fixes the special case we had
for pulumiResourceName and pulumiResourceType.