replace atotto/clipboard with version that lazy loads DLLs (#576)
Similar to https://github.com/pulumi/pulumi/pull/17410, pulumi-dotnet
also uses the atotto/clipboard library, which can sometimes fail when
loading DLLs. Replace this with a version that lazy loads the DLLs,
which solves to avoid the panic when loading the library.

Part of https://github.com/pulumi/pulumi/issues/19185