Simplify `DocLanguageHelper.GetFunctionName` (#19268)
```patch
-	GetFunctionName(modName string, f *schema.Function) string
+	GetFunctionName(f *schema.Function) string
```

The special bookkeeping that `modName` requires is used only within Go's
`DocLanguageHelper`, and `go_gen.DocLanguageHelper` should manage the
bookkeeping as necessary here. This complexity does not need to be
exposed to callers of the function.

If this change is approved, I'll update the `GetFunctionName` method in
pulumi-dotnet and pulumi-yaml.