Detect dangling references in `pulumi schema check` (#19216)
In the case of some larger providers such as AWS, it's helpful to
generate dangling references in the provider schema that we then resolve
with custom code for things that don't fit nicely into the codegen
shapes. Because of this, dangling references aren't considered a problem
with a schema, because they may well be resolved later.

On the other hand, when we have a schema that we expect to be complete,
it's a bit annoying not to be able to check that all types are resolved
just in case we meant for them to be dangling. This PR updates the
behaviour of `pulumi schema check` to forbid dangling references, and
adds a flag, `--allow-dangling-references`, to turn off this particular
warning.

Fixes #17580.

---------

Co-authored-by: Thomas Gummerer <t.gummerer@gmail.com>