implement Plugin debugging for dotnet
Implement plugin debugging for dotnet.  This is unfortunately quite
awkward, because dotnet doesn't have a command line option to say
start the program and attach the program. Instead the program itself
needs to check whether a debugger is attached or not, and wait for it
otherwise.

This PR implements this the same way we do for programs, by setting an
environment variable and then waiting in the provider startup code.
Unfortunately this means it only works for providers written with our
provider framework currently, but not with other types of plugins.