Move Provider.Urn to the core SDK (#619)
We need to start splitting the "Provider" namespace up to also support
policy. One part that will be used by both is the concept of "URN". This
is stable enough that I think we're fine just moving it down into the
core SDK replacing the URN class currently there with a value struct.

---------

Co-authored-by: Justin Van Patten <jvp@justinvp.com>