Rename Host to Engine (#618)
This better describes what this object is, and can match the naming
we're now using in other SDKs.
It's only some parts of the Go SDK that mixed engine and host names
together because we put the implementation of "Log" on the same object
that hosts all the plugins. No need to carry that misnomer to the other
SDKs. This is a client connection to the grpc Engine service.

There was already an internal interface `IEngine` which was a similar
shape except it used the protobuf message types directly. I've deleted
that and moved all uses of it to the new `IEngine` that's using public
request and enum types.

I've left this in the experimental namespace for now, just in case but I
suspect this is close to us calling stable. The internal usage of this
is fine, experimental is unstable API, not unstable implementation.