dont set the stack as the parent if urn is set (#615)
If an engine read is to be performed as the result of deserializing a
Output<SomeResourceType> it throws:

InvalidOperationException("Trying to acquire Deployment.Stack before
'Run' was called.")

Since inside of MLC Construct no Stack is set and everything needs to
have a parent.

The parent is only neede for transformation apply, so I think it safe
that just set this to null here?

I am not sure about the compution and applying of transformations etc
that could all be skiped, but the options that created from
Pulumi.ResourcePackages.TryConstruct anyways dont set anything other
than URN so we can also just leave it as is

---------

Co-authored-by: Jan Kerkenhoff <jan.kerkenhoff.ext@siemens-energy.com>
Co-authored-by: Fraser Waters <frassle@gmail.com>
Co-authored-by: Thomas Gummerer <t.gummerer@gmail.com>