Implement resource hooks in the .Net SDK (#663)
Resource hooks allow programs to execute arbitary code at various points
in the resource lifecycle, such as before or after the
creation/update/deletion of a resource. This change implements resource
hooks in the .Net SDK, extending the existing "callbacks" machinery that
powers transforms, a feature that has a lot of similarities under the
hood.

Hooks are named entities and are registered with Pulumi independently of
resources, being bound separately as part of resource registration. This
is useful for enabling the same hook to be bound to multiple resources,
and necessary for hooking deletes, which are often produced by the
removal of resources from the program. The following example gives a
taste of how hooks feel in .Net:

```csharp
using System;
using System.Threading;
using System.Threading.Tasks;
using Pulumi;

class ResourceHooksStack : Stack
{
    public ResourceHooksStack()
    {
        var u = new Updatable("updatable", new UpdatableArgs
        {
            Value = "step1",
        }, new CustomResourceOptions
        {
            Hooks = {
                BeforeCreate =
                {
                    new("beforeCreate", async (args, cancellationToken) => {
                        Console.WriteLine($"BeforeCreate: value is {args.NewInputs?["value"]}");
                    }),
                },
            },
        });
    }
}

class Program
{
    static Task<int> Main(string[] args) => Deployment.RunAsync<ResourceHooksStack>();
}
```

This change includes integration tests that exercise successful
before/after create/update/delete hooks, as well as failure conditions
in before/after hooks.

Part of pulumi/pulumi#1691