Pick versioned artifacts when publishing (#667)
When publishing artifacts to NuGet, the release script currently just
picks the first package it finds on the filesystem. This works fine when
we know there will be exactly one, but is not robust to e.g. changes to
our build process that forget to clean things up. We can of course keep
alert to changes to our builds that add development artifacts, but we
should probably also just make the script pickier, attempting instead to
find the precise version we have been asked to publish. This change does
just that.