`Makefile` migration: `make all`, `make help`, `make changelog`, `make test_fast`, remove `restore` calls (#662)
A few little changes to get more in line with the `Makefile` proposal.

Personally, the further we get into this, the more it feels to me that
something like [casey/just](https://github.com/casey/just) would be a
far better fit for what we want. Even `mise` has basic task-running
capabilities. Maintaining a `help` command with a useful amount of
documentation using `make` is a big effort, and the compilers we use are
advanced enough that we probably never want a non-`PHONY` target in this
file. We only really need a task runner.