use macos-latest instead of macos-13 (#705)
macos-13 runners will be retired soon, see

https://github.blog/changelog/2025-09-19-github-actions-macos-13-runner-image-is-closing-down/.

The brownout is not supposed to happen until november 4th, but for some
reason jobs are not being acquired right now anyway.

Note that this also switches the architecture to arm64 instead of
x86_64. I'd expect most MacOS users to be on that now anyway, so that's
probably a good thing to test either way.