Prefer `any` over `interface{}` (#20649)
As of Go 1.18+, `interface{}` has an alias: `any`. This change updates
our codebase to prefer the more modern `any`.

This change was primarily made by running `gofmt -r 'interface{} ->
any'` with some manual search/replaces to update
`sdk/go/pulumi/generate` and comments.

This change does not change the output of sdkgen or programgen. That can
be done as a follow-up.