move go version to 1.24.x; clean up state_move_test (#20642)
While working on https://github.com/pulumi/pulumi/pull/20639, I noticed
that we could clean up the `chdir` if we moved to Go 1.24. We're already
building our CI jobs with 1.24 and 1.25 only, and according to our
release policy we're free to upgrade, as long as we support the last two
LTS releases, and there's a use for new features introduced in the older
of those.

Make that upgrade now, and clean up state_move_test.go using `t.Chdir`
instead of a hand-rolled function. Also use `t.Context()` instead of
`context.Background() in the tests where we can.