Expect failures, fix conformance tests (#716)
It looks like the conformance test failures are because the test data
doesn't exist (all the errors are some variation of "could not open file
or directory"). This PR sets them as expected failures and thus fixes
the conformance test runs for PRs. We should probably generate the test
data at some point, though.