Fix changelog job (#721)
A little embarrassing: the `check-changelog` job has in fact never
worked due to a syntax error in the Bash[^1]. I fixed it this week, and
now it's breaking the deploy because the working syntax has a bug: it
will always find `.gitkeep` in the unreleased changes directory, and
therefore `ls -A` will never be empty. This PR fixes this by noting that
changes are never dot-prefixed, and thus removing the `-A` flag is
enough.

[^1]: Here's an example:
https://github.com/pulumi/pulumi-dotnet/actions/runs/18226052479/job/51897500373#step:4:13