Add source artifacts to our build (#719)
In addition to adding the source artifacts I refactored the command
output and how it is printed in case of error. When doing local
development it was failing but all of the output was locked away on the
process actually running the language runtime. The only return was "Exit
value 1" or some such.

By making this change (which I would recommend all the language runtime
follow) I was able to quickly and easily see the actual problem.

Tested by observing that after building we now see both expected
artifacts output. We needed to manually ask for .snupkg as it will
default to using the older "symbols.nupkg" format otherwise.


 Pulumi.Random.4.0.0-alpha.0.nupkg	Pulumi.Random.4.0.0-alpha.0.snupkg